requirements.txt
# Web Framework
flask>=2.0.0
flask-socketio>=5.0.0
dash>=2.0.0
plotly>=5.0.0

# Data Processing & Analysis
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.1.70
alpha_vantage>=2.3.1
finnhub-python>=2.4.0
ta>=0.10.0  # Technical Analysis library

# Optional Dependencies
# Note: ta-lib requires system-level installation:
# Windows: pip install TA-Lib-binary
# Linux: apt-get install ta-lib
# macOS: brew install ta-lib

# News API
newsapi-python>=0.2.6
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
html5lib>=1.1

# NLP & Machine Learning
spacy>=3.0.0
textblob>=0.15.3
nltk>=3.6.0

# Deep Learning & Financial Analysis
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
pandas-ta==0.3.14b0  # Extended Technical Analysis

# Database
SQLAlchemy>=1.4.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
python-dateutil>=2.8.2
